Add debug to makefile, get lexemes building proper

+++ src/compiler/types/lexeme.h
@@ -2,6 +2,7 @@
 #define LEXEME
 #include <stdbool.h>
 #include <stdlib.h>
+#include "../debug/logger.h"
 #include "../extensions/collection_extensions.h"
 #include "../extensions/memory_extensions.h"
 #include "../extensions/string_extensions.h"
@@ -14,14 +15,14 @@ struct lexeme
 {
     char** tokens;
     char** expressions;
-    unsigned int line_number;
+    size_t line_number;
 };
 
 /// @brief Creates and returns a new `lexeme`. If tokens and expressions are both `NULL`, will return `NULL`.
 /// @param tokens An optional collection of keywords, variables, or operators.
 /// @param expressions An optional collection of expressions which aren't transformed before being invoked.
 /// @param line_number The line number the `lexeme` was written on in the L# source code.
-lexeme* create_lexeme(char** tokens, char** expressions, unsigned int line_number);
+lexeme* create_lexeme(char** tokens, char** expressions, size_t line_number);
 
 /// @brief Deallocates the memory of the `lexeme`.
 void free_lexeme(lexeme* lexeme);
